Modificaciones en UserService, insertando las funciones necesarias para UserController

+++ src/main/java/org/project/expressart/Usuario/domain/UserService.java
@@ -1,32 +1,38 @@
 package org.project.expressart.Usuario.domain;
 
 
+import jakarta.persistence.EntityNotFoundException;
 import jakarta.transaction.Transactional;
 import org.apache.coyote.BadRequestException;
-import org.project.expressart.PerfilArtista.domain.PerfilArtista;
-import org.project.expressart.Usuario.dto.UsuarioRequestDTO;
-import org.project.expressart.Usuario.dto.UsuarioResponseDTO;
+import org.modelmapper.ModelMapper;
+import org.project.expressart.Usuario.dto.UserRequestDTO;
+import org.project.expressart.Usuario.dto.UserResponseDTO;
 import org.project.expressart.Usuario.infrastructure.UsuarioRepository;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+
+import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 public class UserService{
     @Autowired
     private final UsuarioRepository userRepository;
-
+    @Autowired
+    private ModelMapper modelMapper;
 
     public UserService(UsuarioRepository userRepo){
         this.userRepository = userRepo;
     }
 
-    public Page<UsuarioResponseDTO> getAllUsers(Pageable pageable) {
-        return userRepository.findAll(pageable);
+    public List<UserResponseDTO> findAll() {
+        Pageable pageable = PageRequest.of(0, 10);
+        return userRepository.findAllBy(pageable);
     }
 
-    public void createUser(UsuarioRequestDTO userdto) throws BadRequestException {
+    public UserResponseDTO create(UserRequestDTO userdto) throws BadRequestException {
         if (userRepository.existsByEmail(userdto.getEmail()))
             throw new BadRequestException("El email ya está en uso");
         if (userRepository.existsByUsername(userdto.getUsername()))
@@ -40,38 +46,52 @@ public class UserService{
         user.setBiography(userdto.getBiography());
         user.setIsActive(true);
         userRepository.save(user);
+        return modelMapper.map(user, UserResponseDTO.class);
     }
 
-    public UsuarioResponseDTO getUserById(Long id){
+
+    public UserResponseDTO findById(Long id){
         Usuario user = userRepository.findById(id).orElseThrow(()-> new ResourceNotFoundEXception("User not found"));
-        UsuarioResponseDTO dto = modelMapper.map(user, UsuarioResponseDTO.class);
-        return dto;
+        return modelMapper.map(user, UserResponseDTO.class);
     }
 
-    public void deleteUser(Long id) {
+    public UserResponseDTO findByEmail(String email){
+        Usuario user = userRepository.findByEmail(email).orElseThrow(()-> new ResourceNotFoundEXception("User not found"));
+        return modelMapper.map(user, UserResponseDTO.class);
+    }
 
-        userRepository.deleteById(id);
+    public List<UserResponseDTO> findSeguidoresByArtistaId(Long id) {
+        List<Usuario> seguidores = userRepository.findSeguidoresByArtistaId(id);
+        return seguidores.stream()
+                .map(seguidor -> modelMapper.map(seguidor, UserResponseDTO.class))
+                .collect(Collectors.toList());
+    }
+
+    public void delete(Long id) {
+        if (userRepository.existsById(id))
+            userRepository.deleteById(id);
+         else
+            throw new EntityNotFoundException("User with ID " + id + " doesn't exist");
     }
 
     @Transactional
-    public void updateUser(Long id, Usuario user) {
+    public UserResponseDTO update(Long id, UserRequestDTO request) {
         Usuario updatedUser = userRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("User not found"));
 
-        if (user.getUsername()!= null && !user.getEmail().isEmpty())
-            updatedUser.setUsername(user.getUsername());
+        if (request.getUsername()!= null && !request.getEmail().isEmpty())
+            updatedUser.setUsername(request.getUsername());
 
-        if (user.getEmail() != null && !user.getEmail().isEmpty())
-            updatedUser.setEmail(user.getEmail());
+        if (request.getEmail() != null && !request.getEmail().isEmpty())
+            updatedUser.setEmail(request.getEmail());
 
-        if (user.getName() != null && !user.getName().isEmpty())
-            updatedUser.setName(user.getName());
+        if (request.getName() != null && !request.getName().isEmpty())
+            updatedUser.setName(request.getName());
 
-        if (user.getPassword() != null && !user.getPassword().isEmpty())
-            updatedUser.setPassword(user.getPassword());
-
-        updatedUser.setAvatar_url(user.getAvatar_url());
-        updatedUser.setBiography(user.getBiography());
-        updatedUser.setToken(user.getToken());
+        if (request.getPassword() != null && !request.getPassword().isEmpty())
+            updatedUser.setPassword(request.getPassword());
+        updatedUser.setAvatar_url(request.getAvatar_url());
+        updatedUser.setBiography(request.getBiography());
         userRepository.save(updatedUser);
+        return modelMapper.map(updatedUser, UserResponseDTO.class);
     }
 }
